Divide rootReducer into user & items reducer
