+++ src/actions/action-types.js
@@ -1,4 +1,4 @@
-// @flow
-export const RECEIVE_USER = 'RECEIVE_USER';
 export const REQUEST_AUTH = 'REQUEST_AUTH';
+export const RECEIVE_USER = 'RECEIVE_USER';
 export const AUTH_FAILED = 'AUTH_FAILED';
+export const REQUEST_ITEMS = 'REQUEST_ITEMS';
