+++ src/reducers/index.js
@@ -1,37 +1,5 @@
-import {
-  REQUEST_AUTH,
-  RECEIVE_USER,
-  AUTH_FAILED,
-} from '../actions/action-types';
+import { combineReducers } from 'redux';
+import user from './user';
+import items from './items';
 
-const initialState = {
-  isAuthenticated: false,
-  isFetching: false,
-};
-
-const userReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case REQUEST_AUTH:
-      return {
-        ...state,
-        isFetching: true,
-      };
-    case RECEIVE_USER:
-      return {
-        ...state,
-        isFetching: false,
-        isAuthenticated: true,
-      };
-    case AUTH_FAILED:
-      /* TODO */
-      return {
-        ...state,
-        isFetching: false,
-        isAuthenticated: false,
-      };
-    default:
-      return state;
-  }
-};
-
-export default userReducer;
+export default combineReducers({ user, items });
